requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2024.1
Django==5.1.1
django-crispy-forms==2.3
djangorestframework==3.15.2
idna==3.10
pillow==10.4.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
